Move logic into library files
