+++ lib/helpers.js
@@ -168,10 +168,36 @@ function warnUnsupportedOptions(unsupportedOptions, definedOptions, logFunction)
   });
 }
 
+function getRuntimeFilenames(runtime, handler) {
+  let files = null;
+  if (runtime.match(/python/)) {
+    files = {
+      handler: `${handler.toString().split('.')[0]}.py`,
+      deps: 'requirements.txt',
+    };
+  } else if (runtime.match(/node/)) {
+    files = {
+      handler: `${handler.toString().split('.')[0]}.js`,
+      deps: 'package.json',
+    };
+  } else if (runtime.match(/ruby/)) {
+    files = {
+      handler: `${handler.toString().split('.')[0]}.rb`,
+      deps: 'Gemfile',
+    };
+  } else {
+    throw new Error(
+      `The runtime ${runtime} is not supported yet`
+    );
+  }
+  return files;
+}
+
 module.exports = {
   warnUnsupportedOptions,
   loadKubeConfig,
   getKubernetesAPIURL,
   getDefaultNamespace,
   getConnectionOptions,
+  getRuntimeFilenames,
 };
