+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "serverless-kubeless",
-  "version": "0.1.12",
+  "version": "0.1.13",
   "description": "This plugin enables support for Kubeless within the [Serverless Framework](https://github.com/serverless).",
   "main": "index.js",
   "directories": {
@@ -41,6 +41,7 @@
     "eslint-plugin-react": "^6.1.1",
     "fs-extra": "^4.0.1",
     "mocha": "^3.4.2",
+    "nock": "^9.0.14",
     "request": "^2.81.0",
     "sinon": "^2.3.6"
   }
